+++ benchmark/src/main/java/io/takari/bpm/benchmark/Linear10DelegateBenchmark.java
@@ -3,17 +3,18 @@ package io.takari.bpm.benchmark;
 import io.takari.bpm.api.ExecutionContext;
 import io.takari.bpm.api.ExecutionException;
 import io.takari.bpm.api.JavaDelegate;
-import io.takari.bpm.model.EndEvent;
-import io.takari.bpm.model.ExpressionType;
-import io.takari.bpm.model.ProcessDefinition;
-import io.takari.bpm.model.SequenceFlow;
-import io.takari.bpm.model.ServiceTask;
-import io.takari.bpm.model.StartEvent;
+import io.takari.bpm.model.*;
 import io.takari.bpm.task.ServiceTaskResolver;
-import java.util.Arrays;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.Scope;
 import org.openjdk.jmh.annotations.State;
+import org.openjdk.jmh.runner.Runner;
+import org.openjdk.jmh.runner.RunnerException;
+import org.openjdk.jmh.runner.options.Options;
+import org.openjdk.jmh.runner.options.OptionsBuilder;
+import org.openjdk.jmh.runner.options.TimeValue;
+
+import java.util.Arrays;
 
 /**
  * Linear process with 10 delegate tasks. The delegate task is an instance of
@@ -22,32 +23,34 @@ import org.openjdk.jmh.annotations.State;
  */
 public class Linear10DelegateBenchmark {
 
+    private static final String EXPR = "${t}";
+
     @State(Scope.Benchmark)
     public static class BenchmarkState extends AbstractBenchmarkState {
 
         public BenchmarkState() {
-            super(new ProcessDefinition("test", Arrays.asList(
+            super(true, false, new ProcessDefinition("test", Arrays.asList(
                     new StartEvent("start"),
                     new SequenceFlow("f1", "start", "t1"),
-                    new ServiceTask("t1", ExpressionType.DELEGATE, "${t}"),
+                    new ServiceTask("t1", ExpressionType.DELEGATE, EXPR),
                     new SequenceFlow("f2", "t1", "t2"),
-                    new ServiceTask("t2", ExpressionType.DELEGATE, "${t}"),
+                    new ServiceTask("t2", ExpressionType.DELEGATE, EXPR),
                     new SequenceFlow("f3", "t2", "t3"),
-                    new ServiceTask("t3", ExpressionType.DELEGATE, "${t}"),
+                    new ServiceTask("t3", ExpressionType.DELEGATE, EXPR),
                     new SequenceFlow("f4", "t3", "t4"),
-                    new ServiceTask("t4", ExpressionType.DELEGATE, "${t}"),
+                    new ServiceTask("t4", ExpressionType.DELEGATE, EXPR),
                     new SequenceFlow("f5", "t4", "t5"),
-                    new ServiceTask("t5", ExpressionType.DELEGATE, "${t}"),
+                    new ServiceTask("t5", ExpressionType.DELEGATE, EXPR),
                     new SequenceFlow("f6", "t5", "t6"),
-                    new ServiceTask("t6", ExpressionType.DELEGATE, "${t}"),
+                    new ServiceTask("t6", ExpressionType.DELEGATE, EXPR),
                     new SequenceFlow("f7", "t6", "t7"),
-                    new ServiceTask("t7", ExpressionType.DELEGATE, "${t}"),
+                    new ServiceTask("t7", ExpressionType.DELEGATE, EXPR),
                     new SequenceFlow("f8", "t7", "t8"),
-                    new ServiceTask("t8", ExpressionType.DELEGATE, "${t}"),
+                    new ServiceTask("t8", ExpressionType.DELEGATE, EXPR),
                     new SequenceFlow("f9", "t8", "t9"),
-                    new ServiceTask("t9", ExpressionType.DELEGATE, "${t}"),
+                    new ServiceTask("t9", ExpressionType.DELEGATE, EXPR),
                     new SequenceFlow("f10", "t9", "t10"),
-                    new ServiceTask("t10", ExpressionType.DELEGATE, "${t}"),
+                    new ServiceTask("t10", ExpressionType.DELEGATE, EXPR),
                     new SequenceFlow("f11", "t10", "end"),
                     new EndEvent("end"))));
             
@@ -70,17 +73,15 @@ public class Linear10DelegateBenchmark {
         }
     }
     
-    /*
     public static void main(String[] args) throws RunnerException {
         Options opt = new OptionsBuilder()
                 .include(Linear10DelegateBenchmark.class.getSimpleName())
                 .forks(0)
-                .warmupIterations(5)
-                .measurementIterations(5)
+                .warmupIterations(15)
+                .measurementIterations(15)
                 .threads(4)
                 .build();
 
         new Runner(opt).run();
     }
-    */
 }