+++ bpm-engine-impl/src/main/java/io/takari/bpm/task/KeyAwareServiceTaskRegistry.java
@@ -1,0 +1,6 @@
+package io.takari.bpm.task;
+
+public interface KeyAwareServiceTaskRegistry extends ServiceTaskRegistry {
+
+    boolean containsKey(String key);
+}